# backend/docs/swagger.yaml
definitions:
  controller.CreateRoleRequest:
    properties:
      name:
        type: string
    type: object
  controller.GetRoleResponse:
    properties:
      role:
        $ref: '#/definitions/entity.Role'
    type: object
  controller.GetRolesResponse:
    properties:
      roles:
        items:
          $ref: '#/definitions/entity.Role'
        type: array
    type: object
  controller.GetUserResponse:
    properties:
      user:
        $ref: '#/definitions/entity.User'
    type: object
  controller.GetUsersResponse:
    properties:
      users:
        items:
          $ref: '#/definitions/entity.User'
        type: array
    type: object
  entity.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  entity.Role:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  entity.User:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
info:
  contact: {}
  description: heysay
  title: heysay
  version: "1.0"
paths:
  /roles:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.GetRolesResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      summary: 全ロール取得 API
      tags:
      - Role
    post:
      consumes:
      - application/json
      parameters:
      - description: ロール作成リクエスト
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      summary: ロール作成 API
      tags:
      - Role
  /roles/{roleId}:
    get:
      consumes:
      - application/json
      parameters:
      - description: ロール ID
        in: path
        name: roleId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.GetRoleResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      summary: ロール取得 API
      tags:
      - Role
  /users:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.GetUsersResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      summary: 全ユーザー取得 API
      tags:
      - User
  /users/{userId}:
    get:
      consumes:
      - application/json
      parameters:
      - description: ユーザー ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.GetUserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      summary: ユーザー取得 API
      tags:
      - User
swagger: "2.0"